Separate internal and community login results
