+++ app/src/community/java/com/twilio/video/app/auth/CommunityAuthenticator.kt
@@ -17,8 +17,8 @@
 package com.twilio.video.app.auth
 
 import android.content.SharedPreferences
-import com.twilio.video.app.auth.LoginResult.CommunityLoginFailureResult
-import com.twilio.video.app.auth.LoginResult.CommunityLoginSuccessResult
+import com.twilio.video.app.auth.CommunityLoginResult.CommunityLoginFailureResult
+import com.twilio.video.app.auth.CommunityLoginResult.CommunityLoginSuccessResult
 import com.twilio.video.app.data.PASSCODE
 import com.twilio.video.app.data.Preferences.DISPLAY_NAME
 import com.twilio.video.app.data.api.AuthServiceException
@@ -55,7 +55,7 @@ class CommunityAuthenticator constructor(
                     CommunityLoginSuccessResult
                 } catch (e: AuthServiceException) {
                     Timber.e(e, "Failed to retrieve token")
-                    CommunityLoginFailureResult(e.error)
+                    CommunityLoginFailureResult(e.error) as LoginResult
                 }
             } else {
                 CommunityLoginFailureResult()
