+++ app/src/community/java/com/twilio/video/app/ui/login/CommunityLoginActivity.java
@@ -32,10 +32,10 @@ import com.google.android.material.textfield.TextInputEditText;
 import com.google.android.material.textfield.TextInputLayout;
 import com.twilio.video.app.R;
 import com.twilio.video.app.auth.Authenticator;
+import com.twilio.video.app.auth.CommunityLoginResult.CommunityLoginFailureResult;
+import com.twilio.video.app.auth.CommunityLoginResult.CommunityLoginSuccessResult;
 import com.twilio.video.app.auth.LoginEvent.CommunityLoginEvent;
 import com.twilio.video.app.auth.LoginResult;
-import com.twilio.video.app.auth.LoginResult.CommunityLoginFailureResult;
-import com.twilio.video.app.auth.LoginResult.CommunityLoginSuccessResult;
 import com.twilio.video.app.base.BaseActivity;
 import com.twilio.video.app.data.api.AuthServiceError;
 import com.twilio.video.app.ui.room.RoomActivity;
