+++ app/src/test/java/com/twilio/video/app/auth/GoogleAuthProviderTest.kt
@@ -13,7 +13,7 @@ import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
-import com.twilio.video.app.auth.LoginResult.GoogleLoginSuccessResult
+import com.twilio.video.app.auth.InternalLoginResult.GoogleLoginSuccessResult
 import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
 import org.hamcrest.CoreMatchers.equalTo
